+++ test/unit/sdk/resource_test.dart
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import 'package:dartastic_opentelemetry/dartastic_opentelemetry.dart';
+import 'package:dartastic_opentelemetry/src/version.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -122,6 +123,21 @@ void main() {
       final keys = merged.attributes.toMap().keys.toList();
 
       expect(keys, equals(['a', 'b', 'c', 'd', 'e', 'f']));
+    });
+
+    test('default resource includes telemetry.sdk.* attributes', () async {
+      final attrs = OTel.defaultResource!.attributes.toList();
+
+      final sdkLanguage =
+          attrs.firstWhere((a) => a.key == 'telemetry.sdk.language');
+      expect(sdkLanguage.value, equals('dart'));
+
+      final sdkName = attrs.firstWhere((a) => a.key == 'telemetry.sdk.name');
+      expect(sdkName.value, equals('opentelemetry'));
+
+      final sdkVersion =
+          attrs.firstWhere((a) => a.key == 'telemetry.sdk.version');
+      expect(sdkVersion.value, equals(packageVersion));
     });
   });
 }